scan: Clarify raster_scan_t types

+++ scan.h
@@ -5,8 +5,8 @@
 struct raster_scan_t {
     uint16_t center_x, center_y;
     uint16_t step_x, step_y;
-    unsigned int size_x, size_y;
-    unsigned int freq;
+    uint16_t size_x, size_y;
+    uint16_t freq;
 };
 
 void raster_scan(struct raster_scan_t *scan);
